requirements.txt
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1